fix: move get all users to user controller
